+++ packages/button/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/button
 
+## 1.0.10
+
+### Patch Changes
+
+- Add separator
+- Updated dependencies
+  - @pikas-ui/icons@1.0.10
+  - @pikas-ui/loader@1.0.10
+  - @pikas-ui/styles@1.0.10
+
 ## 1.0.9
 
 ### Patch Changes