+++ packages/loader/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/loader
 
+## 1.0.10
+
+### Patch Changes
+
+- Add separator
+- Updated dependencies
+  - @pikas-ui/styles@1.0.10
+
 ## 1.0.9
 
 ### Patch Changes